+++ .env
@@ -1,5 +1,6 @@
 API_URL = /api/v1
 secret = Rox-mon-chien
+token = 'jwt-token'
 CONNECTION_STRING = mongodb+srv://josue:Heroes@clusterfish.d2ihs.mongodb.net/Fish-database-dev?retryWrites=true&w=majority
 DB_NAME = Fish-database-dev
 PORT = 3000
